Resolve conv
