+++ Cargo.toml
@@ -60,7 +60,6 @@ darwinia-ecdsa-authority          = { default-features = false, path = "pallet/e
 darwinia-message-gadget           = { default-features = false, path = "pallet/message-gadget" }
 darwinia-message-transact         = { default-features = false, path = "pallet/message-transact" }
 darwinia-precompile-assets        = { default-features = false, path = "precompile/assets" }
-darwinia-precompile-ark-bls12-381 = { default-features = false, path = "precompile/ark-bls12-381" }
 darwinia-precompile-bls12-381     = { default-features = false, path = "precompile/bls12-381" }
 darwinia-precompile-deposit       = { default-features = false, path = "precompile/deposit" }
 darwinia-precompile-staking       = { default-features = false, path = "precompile/staking" }
