--- precompile/ark-bls12-381/Cargo.toml
@@ -1,57 +1,0 @@
-[package]
-authors.workspace = true
-description       = "Arkworks BLS12-381 based precompile for EVM pallet."
-edition.workspace = true
-name              = "darwinia-precompile-ark-bls12-381"
-readme            = "README.md"
-version.workspace = true
-
-[dependencies]
-# crates.io
-sha2 = { version = "0.10.6", default-features = false }
-
-# arkworks
-ark-bls12-381 = { default-features = false, version = "0.4.0", features = ["curve"] }
-ark-serialize = { default-features = false, version = "0.4.1", features = ["derive"]}
-ark-std       = { default-features = false, version = "0.4.0" }
-ark-ec        = { default-features = false, version = "0.4.1" }
-ark-ff        = { default-features = false, version = "0.4.1" }
-
-
-# frontier
-pallet-evm = { workspace = true }
-fp-evm     = { workspace = true }
-
-# moonbeam
-precompile-utils = { workspace = true }
-
-# substrate
-sp-std = { workspace = true }
-
-[dev-dependencies]
-rand = "0.8.5"
-
-[features]
-default = ["std"]
-std = [
-	# crates.io
-	"sha2/std",
-
-	# arkworks
-	"ark-bls12-381/std",
-	"ark-serialize/std",
-	"ark-std/std",
-	"ark-ec/std",
-	"ark-ff/std",
-
-	# frontier
-	"pallet-evm/std",
-	"fp-evm/std",
-
-	# moonbeam
-	"precompile-utils/std",
-
-	# substrate
-	"sp-std/std",
-]
-
