+++ precompile/bls12-381/Cargo.toml
@@ -1,15 +1,26 @@
 [package]
 authors.workspace = true
-description       = "BLS12-381 implementation for EVM pallet."
+description       = "Arkworks BLS12-381 based precompile for EVM pallet."
 edition.workspace = true
 name              = "darwinia-precompile-bls12-381"
 readme            = "README.md"
 version.workspace = true
 
 [dependencies]
+# crates.io
+sha2 = { version = "0.10.6", default-features = false }
+
+# arkworks
+ark-bls12-381 = { default-features = false, version = "0.4.0", features = ["curve"] }
+ark-serialize = { default-features = false, version = "0.4.1", features = ["derive"]}
+ark-std       = { default-features = false, version = "0.4.0" }
+ark-ec        = { default-features = false, version = "0.4.1" }
+ark-ff        = { default-features = false, version = "0.4.1" }
+
+
 # frontier
-fp-evm     = { workspace = true }
 pallet-evm = { workspace = true }
+fp-evm     = { workspace = true }
 
 # moonbeam
 precompile-utils = { workspace = true }
@@ -18,27 +29,24 @@ precompile-utils = { workspace = true }
 sp-std = { workspace = true }
 
 [dev-dependencies]
-# crates.io
-codec      = { package = "parity-scale-codec", workspace = true }
-scale-info = { workspace = true }
-
-# moonbeam
-precompile-utils = { workspace = true, features = ["testing"] }
-
-# substrate
-frame-system     = { workspace = true }
-pallet-balances  = { workspace = true, features = ["std"] }
-pallet-timestamp = { workspace = true, features = ["std"] }
-sp-core          = { workspace = true }
-sp-io            = { workspace = true }
-sp-runtime       = { workspace = true }
+rand = "0.8.5"
 
 [features]
 default = ["std"]
 std = [
+	# crates.io
+	"sha2/std",
+
+	# arkworks
+	"ark-bls12-381/std",
+	"ark-serialize/std",
+	"ark-std/std",
+	"ark-ec/std",
+	"ark-ff/std",
+
 	# frontier
-	"fp-evm/std",
 	"pallet-evm/std",
+	"fp-evm/std",
 
 	# moonbeam
 	"precompile-utils/std",
@@ -46,3 +54,4 @@ std = [
 	# substrate
 	"sp-std/std",
 ]
+
