+++ shell.nix
@@ -5,10 +5,7 @@ let
       rev = "78e723925daf5c9e8d0a1837ec27059e61649cb6";
     });
   nixpkgs = import <nixpkgs> { overlays = [ mozillaOverlay ]; };
-  rust-nightly = with nixpkgs; ((rustChannelOf { date = "2022-11-15"; channel = "nightly"; }).rust.override {
-    extensions = [ "rust-src" ];
-    targets = [ "wasm32-unknown-unknown" ];
-  });
+  rust-nightly = with nixpkgs; (nixpkgs.latest.rustChannels.nightly.rust);
 in
 with nixpkgs; pkgs.mkShell {
   nativeBuildInputs = [
